moved wikidb into db, moved migrations, updated imports
