+++ src/wiki/getdoc.ts
@@ -2,11 +2,11 @@ import { LambdaBuilder } from '../util/middleware/middleware';
 // import { APIGatewayProxyEvent } from 'aws-lambda';
 import { APIResponse, SuccessResponse } from '../util/middleware/response';
 import { InputValidator } from '../util/middleware/inputValidator';
-import { getWikiDatabase } from '../util/wikidb';
+import { getDatabase } from '../util/db';
 import { S3 } from 'aws-sdk';
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda';
 
-const db = getWikiDatabase();
+const db = getDatabase();
 export const handler = new LambdaBuilder(router)
     .use(new InputValidator())
     // Add any additional middleware as needed
@@ -57,7 +57,7 @@ export async function getDoc(docName: string, areaName: string) {
             'lastUpdatedDate',
         ])
         .where('name', '=', docName)
-        .where('id', '=', areaID)
+        .where('areaID', '=', areaID)
         .executeTakeFirst();
 
     if (!docRes) {
